Added Statistics Chart
The Bar Chart in the Statistics Fragment is now being displayed.

The Bar Chart uses MPAndroidChart to simplify rendering.
Base Configuration happens in setupBarChart(), rest happens in prepareAndSetBarChartData() depending on the TimePeriod Type.

The X-Axis shows the week days (day of month), every other day of the month, or every second (odd) month of the year.
This is done to not overfill the X-Axis.

---

The Chart is refreshed every time the Completions themselves are refreshed.