Added Statistics UI Test
Added a Statistics UI Test that tests various things:
- Switching Task Categories
- Switching Periodicties
- Switching Date Ranges

Tests for completion counts, completion names, date range text...

To mock the Calendar in the StatisticsFragment, the Fragment now uses a Calendar Interface.
By default, the "SystemCalendar" class is used, which just falls back to the normal Calender.
During testing, a FakeCalendar is used, which uses a prepopulated Calender to ensure consistency with testing.