DateHelper Fix
Fixed getCurrentPeriodicityRange in DateHelper returning wrong values if the current day is a Sunday (and possibly Monday).
Due to the Week (for this specific class) always starting on a Monday, having the current day be Sunday caused it to return values for the *next* week instead.
Happened due to Sunday being represented as 1.